(failing) test for split-brain scenario

+++ core/src/test/java/brooklyn/entity/rebind/persister/ListeningObjectStore.java
@@ -11,6 +11,7 @@ import org.slf4j.LoggerFactory;
 
 import brooklyn.management.ManagementContext;
 import brooklyn.management.ha.HighAvailabilityMode;
+import brooklyn.util.collections.MutableList;
 import brooklyn.util.text.Strings;
 import brooklyn.util.time.CountdownTimer;
 import brooklyn.util.time.Duration;
@@ -20,7 +21,8 @@ import com.google.common.base.Preconditions;
 public class ListeningObjectStore implements PersistenceObjectStore {
 
     protected final PersistenceObjectStore delegate;
-    protected final ObjectStoreTransactionListener listener;
+    protected final List<ObjectStoreTransactionListener> listeners = MutableList.of();
+    private boolean writesFailSilently = false;
 
     public static interface ObjectStoreTransactionListener {
         public void recordQueryOut(String summary, int size);
@@ -102,9 +104,10 @@ public class ListeningObjectStore implements PersistenceObjectStore {
         }
     }
 
-    public ListeningObjectStore(PersistenceObjectStore delegate, ObjectStoreTransactionListener listener) {
+    public ListeningObjectStore(PersistenceObjectStore delegate, ObjectStoreTransactionListener ...listeners) {
         this.delegate = Preconditions.checkNotNull(delegate);
-        this.listener = Preconditions.checkNotNull(listener);
+        for (ObjectStoreTransactionListener listener: listeners)
+            this.listeners.add(listener);
     }
 
     @Override
@@ -124,15 +127,23 @@ public class ListeningObjectStore implements PersistenceObjectStore {
 
     @Override
     public void createSubPath(String subPath) {
-        listener.recordQueryOut("creating path "+subPath, 1+subPath.length());
+        if (writesFailSilently)
+            return;
+        
+        for (ObjectStoreTransactionListener listener: listeners)
+            listener.recordQueryOut("creating path "+subPath, 1+subPath.length());
         delegate.createSubPath(subPath);
     }
 
     @Override
     public List<String> listContentsWithSubPath(String subPath) {
-        listener.recordQueryOut("requesting list "+subPath, 1+subPath.length());
+        for (ObjectStoreTransactionListener listener: listeners)
+            listener.recordQueryOut("requesting list "+subPath, 1+subPath.length());
+        
         List<String> result = delegate.listContentsWithSubPath(subPath);
-        listener.recordDataIn("receiving list "+subPath, result.toString().length());
+        
+        for (ObjectStoreTransactionListener listener: listeners)
+            listener.recordDataIn("receiving list "+subPath, result.toString().length());
         return result;
     }
 
@@ -153,7 +164,8 @@ public class ListeningObjectStore implements PersistenceObjectStore {
 
     @Override
     public void deleteCompletely() {
-        listener.recordDataOut("deleting completely", 1);
+        for (ObjectStoreTransactionListener listener: listeners)
+            listener.recordDataOut("deleting completely", 1);
         delegate.deleteCompletely();
     }
 
@@ -172,24 +184,39 @@ public class ListeningObjectStore implements PersistenceObjectStore {
         }
         @Override
         public void put(String val) {
-            listener.recordDataOut("writing "+path, val.length());
+            if (writesFailSilently)
+                return;
+
+            for (ObjectStoreTransactionListener listener: listeners)
+                listener.recordDataOut("writing "+path, val.length());
             delegate.put(val);
         }
         @Override
         public void append(String s) {
-            listener.recordDataOut("appending "+path, s.length());
+            if (writesFailSilently)
+                return;
+
+            for (ObjectStoreTransactionListener listener: listeners)
+                listener.recordDataOut("appending "+path, s.length());
             delegate.append(s);
         }
         @Override
         public void delete() {
-            listener.recordQueryOut("deleting "+path, path.length());
+            if (writesFailSilently)
+                return;
+
+            for (ObjectStoreTransactionListener listener: listeners)
+                listener.recordQueryOut("deleting "+path, path.length());
             delegate.delete();
         }
         @Override
         public String get() {
-            listener.recordQueryOut("requesting "+path, path.length());
+            for (ObjectStoreTransactionListener listener: listeners)
+                listener.recordQueryOut("requesting "+path, path.length());
             String result = delegate.get();
-            listener.recordDataIn("reading "+path, result.length());
+            
+            for (ObjectStoreTransactionListener listener: listeners)
+                listener.recordDataIn("reading "+path, (result==null ? 0 : result.length()));
             return result;
         }
         @Override
@@ -198,4 +225,7 @@ public class ListeningObjectStore implements PersistenceObjectStore {
         }
     }
 
+    public void setWritesFailSilently(boolean writesFailSilently) {
+        this.writesFailSilently = writesFailSilently;
+    }
 }
